Docker_bdm/requirements_airflow.txt

apache-airflow
apache-airflow-providers-apache-spark
apache-airflow-providers-common-compat
psycopg2-binary
sqlalchemy

geopandas
folium>=0.12
matplotlib
mapclassify
pandas
numpy
requests
validators
email-validator
deep-translator
swifter
chardet
beautifulsoup4
python-dotenv
azure-storage-blob
azure-storage-file-datalake
hdfs
pymilvus
geopy
tqdm
pillow
findspark